RAG-upload/deprecated/backup/assets/requirements.txt
python-dotenv
pinecone
pinecone-client
pinecone-plugin-assistant